+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2026-04-09
+
+### Added
+- `--version` command to display project name, version, author, and website.
+- Official documentation link to `https://xxcc-unicorn-press.github.io`.
+
+### Changed
+- Updated project author and copyright to `XXCC Unicorn Press Editorial Board`.
+
 ## [0.2.0] - 2026-04-09
 
 ### Added